Revise the full pipeline.
The meta-encoding for the derivation sequence is applied several times to assign atoms in stage (as there is no clear way to have the order of derivation) and the answer set is also filtered (just in case the same atom is derived with multiple reasons on the same iteration).

The meta-encoding for the graph is aligned with the definition of "reason" in the article.

Disjunctive heads are supported.